requirements.txt
python-telegram-bot >= 21
aiohttp
aiosqlite >= 0.20
dacite
pyyaml
aiofiles
